+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/search/ActionSearchResult.java
@@ -31,7 +31,7 @@ import java.util.Map;
  *
  * @author JPEXS
  */
-public class ActionSearchResult {
+public class ActionSearchResult implements ScriptSearchResult {
 
     private final ASMSource src;