+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/search/ScriptSearchResult.java
@@ -1,0 +1,9 @@
+package com.jpexs.decompiler.flash.search;
+
+/**
+ *
+ * @author JPEXS
+ */
+public interface ScriptSearchResult {
+
+}